drm/i915: Introduce the HWCONFIG header
GuC contains a consolidated table with all needed information for the
device in use.

Right now this information is spread and hardcoded to all the components
including GuC, i915 and all user space components. The goal here is
to consolidate the definition into GuC in a way that all interested
components can grab the very latest and synchronized information.

A host to GuC interface provides a mechanism for i915 to allocate
a pinned buffer on GGTT and cache it through the entire runtime period.

Then a query UAPI will be added in a way that user space components
can grab the current table in a blob format at any time.

This header is a separated change so it can be easily seen and copied
to the user space components and also to be flexible during initial
development.

Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Signed-off-by: Ramalingam C <ramalingam.c@intel.com>