drm/i915/uapi: Extend query ioctl to return hwconfig table
GuC contains a consolidated table with all needed information for the
device in use.

Right now this information is spread and hardcoded to all the components
including GuC, i915 and all user space components. The goal here is
to consolidate the definition into GuC in a way that all interested
components can grab the very latest and synchronized information using
this simple query.

As most of other queries, this one can be called twice,
one with item.length=0, so it can determine the exact buffer size,
then allocate the user memory and call it again for grabbing the
table. Simple as this:
  memset(&item, 0, sizeof(item));
  item.query_id = DRM_I915_QUERY_HWCONCFIG_TABLE;
  query.items_ptr = (int64_t) &item;
  query.num_items = 1;

  ioctl(fd, DRM_IOCTL_I915_QUERY, query, sizeof(query));

  if (item.size <= 0)
    return -ENOENT;

  data = malloc(item.length);
  item.data_ptr = (int64_t) &data;
  ioctl(fd, DRM_IOCTL_I915_QUERY, query, sizeof(query));

  // Parses the data accordingly ...

The returned array is a simple and flexible KLV (Key/Length/Value)
formatted table. For instance it could be simple as this:

  enum device_attr {
     ATTR_EUS_PER_SSLICE = 0,
     ATTR_SOME_MASK    = 1,
  };

  static const u32 hwconfig[] = {
      ATTR_EUS_PER_SSLICE,
      1,             // Value Length in DWords
      8,             // Value

      ATTR_SOME_MASK,
      3,
      0x00FFFFFFFF, 0xFFFFFFFF, 0xFF000000, // Value
  };
  static const u32 table_size = sizeof(hwconfig) / sizeof(hwconfig[0]));

It is important to highlight though that the device attribute ids are
common across multiple components including GuC, i915 and user space
components. The definition of the actual and current attributes can be
found in the header file: intel_guc_hwconfig_types.h

Cc: Tvrtko Ursulin <tvrtko.ursulin@linux.intel.com>
Cc: Kenneth Graunke <kenneth.w.graunke@intel.com>
Cc: Michal Wajdeczko <michal.wajdeczko@intel.com>
Cc: Slawomir Milczarek <slawomir.milczarek@intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Signed-off-by: Ramalingam C <ramalingam.c@intel.com>